requirements.txt
cassandra-driver>=3.13
six>=1.6
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
djangorestframework==3.12.2
django-utils-six
djangorestframework-simplejwt==4.4.0
idna==2.10
Pillow==8.0.1
PyJWT==1.7.1
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
time-uuid

